+++ plugins/source/aws/client/mocks/quicksight.go
@@ -1745,6 +1745,36 @@ func (mr *MockQuicksightClientMockRecorder) ListIAMPolicyAssignmentsForUser(arg0
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListIAMPolicyAssignmentsForUser", reflect.TypeOf((*MockQuicksightClient)(nil).ListIAMPolicyAssignmentsForUser), varargs...)
 }
 
+// ListIdentityPropagationConfigs mocks base method.
+func (m *MockQuicksightClient) ListIdentityPropagationConfigs(arg0 context.Context, arg1 *quicksight.ListIdentityPropagationConfigsInput, arg2 ...func(*quicksight.Options)) (*quicksight.ListIdentityPropagationConfigsOutput, error) {
+
+	// Assertion inserted by client/mockgen/main.go
+	o := &quicksight.Options{}
+	for _, f := range arg2 {
+		f(o)
+	}
+	if o.Region == "" {
+		m.ctrl.T.Errorf("Region not set in call to ListIdentityPropagationConfigs")
+	}
+
+	m.ctrl.T.Helper()
+	varargs := []interface{}{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ListIdentityPropagationConfigs", varargs...)
+	ret0, _ := ret[0].(*quicksight.ListIdentityPropagationConfigsOutput)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ListIdentityPropagationConfigs indicates an expected call of ListIdentityPropagationConfigs.
+func (mr *MockQuicksightClientMockRecorder) ListIdentityPropagationConfigs(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListIdentityPropagationConfigs", reflect.TypeOf((*MockQuicksightClient)(nil).ListIdentityPropagationConfigs), varargs...)
+}
+
 // ListIngestions mocks base method.
 func (m *MockQuicksightClient) ListIngestions(arg0 context.Context, arg1 *quicksight.ListIngestionsInput, arg2 ...func(*quicksight.Options)) (*quicksight.ListIngestionsOutput, error) {