+++ plugins/source/aws/client/transformers.go
@@ -58,3 +58,11 @@ func TimestampResolverTransformer(field reflect.StructField, path string) schema
 		return r.Set(c.Name, t)
 	}
 }
+
+// TagsResolverTransformer adds possibility to override the column resolver for "Tags" field
+func TagsResolverTransformer(field reflect.StructField, path string) schema.ColumnResolver {
+	if path == "Tags" {
+		return ResolveTags
+	}
+	return transformers.DefaultResolverTransformer(field, path)
+}