+++ plugins/source/aws/resources/services/docdb/event_subscriptions.go
@@ -3,6 +3,7 @@ package docdb
 import (
 	"context"
 
+	"github.com/apache/arrow/go/v14/arrow"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/docdb"
 	"github.com/aws/aws-sdk-go-v2/service/docdb/types"
@@ -22,6 +23,12 @@ func EventSubscriptions() *schema.Table {
 		Columns: []schema.Column{
 			client.DefaultAccountIDColumn(false),
 			client.DefaultRegionColumn(false),
+			{
+				Name:       "arn",
+				Type:       arrow.BinaryTypes.String,
+				Resolver:   schema.PathResolver("EventSubscriptionArn"),
+				PrimaryKey: true,
+			},
 		},
 	}
 }