+++ plugins/source/aws/resources/services/elasticbeanstalk/environments.go
@@ -22,23 +22,23 @@ func Environments() *schema.Table {
 		Multiplex:   client.ServiceAccountRegionMultiplexer(tableName, "elasticbeanstalk"),
 		Transform:   transformers.TransformWithStruct(&types.EnvironmentDescription{}),
 		Columns: []schema.Column{
-			client.DefaultAccountIDColumn(true),
-			{
-				Name:     "arn",
-				Type:     arrow.BinaryTypes.String,
-				Resolver: schema.PathResolver("EnvironmentArn"),
-			},
+			client.DefaultAccountIDColumn(false),
 			client.DefaultRegionColumn(false),
+			{
+				Name:       "arn",
+				Type:       arrow.BinaryTypes.String,
+				Resolver:   schema.PathResolver("EnvironmentArn"),
+				PrimaryKey: true,
+			},
 			{
 				Name:     "tags",
 				Type:     sdkTypes.ExtensionTypes.JSON,
 				Resolver: resolveElasticbeanstalkEnvironmentTags,
 			},
 			{
-				Name:       "id",
-				Type:       arrow.BinaryTypes.String,
-				Resolver:   schema.PathResolver("EnvironmentId"),
-				PrimaryKey: true,
+				Name:     "id",
+				Type:     arrow.BinaryTypes.String,
+				Resolver: schema.PathResolver("EnvironmentId"),
 			},
 			{
 				Name:     "listeners",