+++ plugins/source/aws/resources/services/glue/datacatalog_encryption_settings.go
@@ -25,7 +25,7 @@ func DatacatalogEncryptionSettings() *schema.Table {
 	}
 }
 
-func fetchGlueDatacatalogEncryptionSettings(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+func fetchGlueDatacatalogEncryptionSettings(ctx context.Context, meta schema.ClientMeta, _ *schema.Resource, res chan<- any) error {
 	cl := meta.(*client.Client)
 	svc := cl.Services(client.AWSServiceGlue).Glue
 	result, err := svc.GetDataCatalogEncryptionSettings(ctx, &glue.GetDataCatalogEncryptionSettingsInput{}, func(options *glue.Options) {