+++ plugins/source/aws/resources/services/lambda/layer_version_policies.go
@@ -14,14 +14,18 @@ func layerVersionPolicies() *schema.Table {
 		Name:        tableName,
 		Description: `https://docs.aws.amazon.com/lambda/latest/dg/API_GetLayerVersionPolicy.html`,
 		Resolver:    fetchLambdaLayerVersionPolicies,
-		Transform:   transformers.TransformWithStruct(&lambda.GetLayerVersionPolicyOutput{}, transformers.WithSkipFields("ResultMetadata")),
+		Transform: transformers.TransformWithStruct(&lambda.GetLayerVersionPolicyOutput{},
+			transformers.WithPrimaryKeys("RevisionId"),
+			transformers.WithSkipFields("ResultMetadata"),
+		),
 		Columns: []schema.Column{
 			client.DefaultAccountIDColumn(false),
 			client.DefaultRegionColumn(false),
 			{
-				Name:     "layer_version_arn",
-				Type:     arrow.BinaryTypes.String,
-				Resolver: schema.ParentColumnResolver("arn"),
+				Name:       "layer_version_arn",
+				Type:       arrow.BinaryTypes.String,
+				Resolver:   schema.ParentColumnResolver("arn"),
+				PrimaryKey: true,
 			},
 			{
 				Name:     "layer_version",