+++ plugins/source/aws/resources/services/rds/db_parameter_groups.go
@@ -37,7 +37,7 @@ func DbParameterGroups() *schema.Table {
 		},
 
 		Relations: []*schema.Table{
-			DbParameterGroupDbParameters(),
+			dbParameterGroupDbParameters(),
 		},
 	}
 }